backend-ai/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==3.2.2
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
fastapi==0.104.1
google-generativeai==0.8.4
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
pydantic==2.5.0
pydantic_core==2.14.1
pymongo==4.16.0
python-dotenv==1.2.1
requests==2.31.0
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
watchfiles==1.1.1
websockets==16.0
